Blocks drop their cooresponding item entities and id enums were refactored to avoid circular imports
