+++ scripts/itemData.js
@@ -4,35 +4,10 @@
 // =======================================================================
 // Importing
 
-import { BlockId } from "./blockData.js";
+import { BlockId } from "./blockId.js";
+import { ItemId } from "./itemId.js";
 
 // =======================================================================
-
-// Enum of item IDs
-export const ItemId = {
-    AirBlock:           0,
-    GrassBlock:         1,
-    DirtBlock:          2,
-    StoneBlock:         3,
-    SandBlock:          4,
-    WaterBlock:         5,
-    LogBlock:           6,
-    LeavesBlock:        7,
-    CoalOreBlock:       8,
-    IronOreBlock:       9,
-    GoldOreBlock:       10,
-    DiamondOreBlock:    11,
-    AridGrassBlock:     12,
-    TropicalGrassBlock: 13,
-    ColdGrassBlock:     14,
-    CactusBlock:        15,
-    JungleLogBlock:     16,
-    PineLogBlock:       17,
-    AcaciaLogBlock:     18,
-    JungleLeavesBlock:  19,
-    PineLeavesBlock:    20,
-    AcaciaLeavesBlock:  21,
-};
 
 // Note: order must match enum order
 export const itemStaticData = [
