+++ scripts/itemId.js
@@ -1,0 +1,34 @@
+// Minecraft clone made with THREE.js
+// Item Ids
+// Note that this is separate from itemData to avoid circular imports
+// By Amy Burnett
+// =======================================================================
+// Importing
+
+// =======================================================================
+
+// Enum of item IDs
+export const ItemId = {
+    AirBlock:           0,
+    GrassBlock:         1,
+    DirtBlock:          2,
+    StoneBlock:         3,
+    SandBlock:          4,
+    WaterBlock:         5,
+    LogBlock:           6,
+    LeavesBlock:        7,
+    CoalOreBlock:       8,
+    IronOreBlock:       9,
+    GoldOreBlock:       10,
+    DiamondOreBlock:    11,
+    AridGrassBlock:     12,
+    TropicalGrassBlock: 13,
+    ColdGrassBlock:     14,
+    CactusBlock:        15,
+    JungleLogBlock:     16,
+    PineLogBlock:       17,
+    AcaciaLogBlock:     18,
+    JungleLeavesBlock:  19,
+    PineLeavesBlock:    20,
+    AcaciaLeavesBlock:  21,
+};
